#2225 :truck: move mbw component files in ergebnismeldung/MBW
