+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/MBW/stapelAB/TheMBWGueltigeStimmenAnzeigenTable.vue
