+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/MBW/stapelAB/TheMBWWaehlerAnzeigenCard.vue
