+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/MBW/stapelAB/TheMBWWahlberechtigteAnzeigenTable.vue
