+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/MBW/stapelBC/BaseRowKandidatStimmenErfassen.vue
