+++ wls-gui-wahllokalsystem/src/components/ergebnismeldung/MBW/stapelC/TheMBWUngueltigeStimmenAnzeigenCard.vue
