+++ wls-gui-wahllokalsystem/src/views/ergebnismeldung/MBW/MBWNiederschriftView.vue
